Consolidation of TelemetryUtils into TelemetryManagementInterface
