+++ clients/src/main/java/org/apache/kafka/clients/producer/KafkaProducer.java
@@ -38,7 +38,6 @@ import org.apache.kafka.clients.producer.internals.RecordAccumulator;
 import org.apache.kafka.clients.producer.internals.Sender;
 import org.apache.kafka.clients.producer.internals.TransactionManager;
 import org.apache.kafka.clients.producer.internals.TransactionalRequestResult;
-import org.apache.kafka.clients.telemetry.TelemetryUtils;
 import org.apache.kafka.common.Cluster;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.Metric;
@@ -365,7 +364,7 @@ public class KafkaProducer<K, V> implements Producer<K, V> {
                     config.originalsWithPrefix(CommonClientConfigs.METRICS_CONTEXT_PREFIX));
             this.metrics = new Metrics(metricConfig, reporters, time, metricsContext);
             this.producerMetrics = new KafkaProducerMetrics(metrics);
-            this.tmi = TelemetryUtils.maybeCreateTmi(config, time, clientId);
+            this.tmi = TelemetryManagementInterface.maybeCreate(config, time, clientId);
 
             if (this.tmi != null) {
                 this.producerTelemetryRegistry = new ProducerTelemetryRegistry(tmi.metrics());
