+++ clients/src/main/java/org/apache/kafka/clients/telemetry/TelemetryManagementInterface.java
@@ -16,25 +16,46 @@
  */
 package org.apache.kafka.clients.telemetry;
 
+import static org.apache.kafka.common.Uuid.ZERO_UUID;
+
 import java.io.Closeable;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.nio.ByteBuffer;
 import java.time.Duration;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
+import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.common.KafkaException;
+import org.apache.kafka.common.MetricName;
 import org.apache.kafka.common.Uuid;
+import org.apache.kafka.common.config.AbstractConfig;
 import org.apache.kafka.common.errors.InterruptException;
+import org.apache.kafka.common.message.GetTelemetrySubscriptionsResponseData;
+import org.apache.kafka.common.metrics.Gauge;
+import org.apache.kafka.common.metrics.KafkaMetric;
 import org.apache.kafka.common.metrics.KafkaMetricsContext;
 import org.apache.kafka.common.metrics.Measurable;
 import org.apache.kafka.common.metrics.MetricConfig;
 import org.apache.kafka.common.metrics.Metrics;
 import org.apache.kafka.common.metrics.MetricsContext;
 import org.apache.kafka.common.metrics.stats.CumulativeSum;
+import org.apache.kafka.common.metrics.stats.Histogram;
 import org.apache.kafka.common.record.CompressionType;
+import org.apache.kafka.common.record.RecordBatch;
+import org.apache.kafka.common.requests.AbstractRequest;
+import org.apache.kafka.common.requests.GetTelemetrySubscriptionRequest;
+import org.apache.kafka.common.requests.PushTelemetryRequest;
+import org.apache.kafka.common.utils.ByteBufferOutputStream;
+import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.common.utils.Time;
+import org.apache.kafka.common.utils.Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -117,7 +138,7 @@ public class TelemetryManagementInterface implements Closeable {
         }
     }
 
-    public ByteBuffer collectMetricsPayload(CompressionType compressionType, boolean deltaTemporality) throws IOException {
+    public Bytes collectMetricsPayload(CompressionType compressionType, boolean deltaTemporality) {
         log.trace("collectMetricsPayload starting");
         Collection<TelemetryMetric> telemetryMetrics = telemetryMetricsReporter.current().stream().map(kafkaMetric -> {
             String name = kafkaMetric.metricName().name();
@@ -135,7 +156,7 @@ public class TelemetryManagementInterface implements Closeable {
                 }
             }
 
-            MetricType metricType = TelemetryUtils.metricType(kafkaMetric);
+            MetricType metricType = metricType(kafkaMetric);
             String description = kafkaMetric.metricName().description();
 
             TelemetryMetric telemetryMetric = new TelemetryMetric(name, metricType, value, description);
@@ -143,7 +164,8 @@ public class TelemetryManagementInterface implements Closeable {
             return telemetryMetric;
         }).collect(Collectors.toList());
 
-        return TelemetryUtils.serialize(telemetryMetrics, compressionType, telemetrySerializer);
+        ByteBuffer buf = serialize(telemetryMetrics, compressionType, telemetrySerializer);
+        return Bytes.wrap(Utils.readBytes(buf));
     }
 
     public void setSubscription(TelemetrySubscription subscription) {
@@ -212,7 +234,7 @@ public class TelemetryManagementInterface implements Closeable {
     public void setState(TelemetryState state) {
         synchronized (stateLock) {
             log.trace("Setting state from {} to {}", this.state, state);
-            this.state = this.state.validateTransition(state);
+            this.state = validateTransition(this.state, state);
         }
     }
 
@@ -220,6 +242,329 @@ public class TelemetryManagementInterface implements Closeable {
         synchronized (stateLock) {
             return state;
         }
+    }
+
+    public void telemetrySubscriptionFailed(Throwable error) {
+        if (error != null)
+            log.warn("Failed to retrieve telemetry subscription; using existing subscription", error);
+        else
+            log.warn("Failed to retrieve telemetry subscription; using existing subscription");
+
+        setState(TelemetryState.subscription_needed);
+    }
+
+    public void pushTelemetryFailed(Exception error) {
+        if (error != null)
+            log.warn("Failed to push telemetry", error);
+        else
+            log.warn("Failed to push telemetry", new Exception());
+
+        TelemetryState state = state();
+
+        if (state == TelemetryState.push_in_progress)
+            setState(TelemetryState.subscription_needed);
+        else if (state == TelemetryState.terminating_push_in_progress)
+            setState(TelemetryState.terminated);
+        else
+            throw new IllegalTelemetryStateException(String.format("Could not transition state after failed push telemetry from state %s", state));
+    }
+
+    public void telemetrySubscriptionSucceeded(GetTelemetrySubscriptionsResponseData data) {
+        Time time = time();
+        Set<MetricName> metricNames = validateMetricNames(data.requestedMetrics());
+        List<CompressionType> acceptedCompressionTypes = validateAcceptedCompressionTypes(data.acceptedCompressionTypes());
+        Uuid clientInstanceId = validateClientInstanceId(data.clientInstanceId());
+        int pushIntervalMs = validatePushInteravlMs(data.pushIntervalMs());
+
+        TelemetrySubscription telemetrySubscription = new TelemetrySubscription(time.milliseconds(),
+            data.throttleTimeMs(),
+            clientInstanceId,
+            data.subscriptionId(),
+            acceptedCompressionTypes,
+            pushIntervalMs,
+            data.deltaTemporality(),
+            metricNames);
+
+        log.debug("Successfully retrieved telemetry subscription: {}", telemetrySubscription);
+
+        setSubscription(telemetrySubscription);
+        setState(TelemetryState.push_needed);
+    }
+
+    public void pushTelemetrySucceeded() {
+        log.trace("Successfully pushed telemetry");
+
+        TelemetryState state = state();
+
+        if (state == TelemetryState.push_in_progress)
+            setState(TelemetryState.subscription_needed);
+        else if (state == TelemetryState.terminating_push_in_progress)
+            setState(TelemetryState.terminated);
+        else
+            throw new IllegalTelemetryStateException(String.format("Could not transition state after successful push telemetry from state %s", state));
+    }
+
+    public boolean isNetworkState() {
+        TelemetryState s = state();
+        return s == TelemetryState.subscription_in_progress || s == TelemetryState.push_in_progress || s == TelemetryState.terminating_push_in_progress;
+    }
+
+    public boolean isTerminatedState() {
+        TelemetryState s = state();
+        return s == TelemetryState.terminated;
+    }
+
+    public long timeToNextUpdate() {
+        long t = 0;
+        TelemetryState s = state();
+
+        if (s == TelemetryState.initialized || s == TelemetryState.subscription_needed) {
+            // TODO: TELEMETRY_TODO: verify
+            t = 0;
+        } else  if (s == TelemetryState.terminated) {
+            // TODO: TELEMETRY_TODO: verify and add a good error message
+            throw new IllegalTelemetryStateException();
+        } else {
+            long milliseconds = time().milliseconds();
+            TelemetrySubscription subscription = subscription();
+
+            if (subscription != null) {
+                long fetchMs = subscription.fetchMs();
+                long pushIntervalMs = subscription.pushIntervalMs();
+                t = fetchMs + pushIntervalMs - milliseconds;
+
+                if (t < 0)
+                    t = 0;
+            }
+        }
+
+        log.debug("For state {}, returning {} for time to next update", s, t);
+        return t;
+    }
+
+    public AbstractRequest.Builder<?> createRequest() {
+        TelemetrySubscription subscription = subscription();
+
+        if (state() == TelemetryState.subscription_needed) {
+            setState(TelemetryState.subscription_in_progress);
+            Uuid clientInstanceId = subscription != null ? subscription.clientInstanceId() : ZERO_UUID;
+            return new GetTelemetrySubscriptionRequest.Builder(clientInstanceId);
+        } else if (state() == TelemetryState.push_needed) {
+            if (subscription == null)
+                throw new IllegalStateException(String.format("Telemetry state is %s but subscription is null", state()));
+
+            boolean terminating = state() == TelemetryState.terminating;
+            CompressionType compressionType = preferredCompressionType(subscription.acceptedCompressionTypes());
+            Bytes metricsData = collectMetricsPayload(compressionType, subscription.deltaTemporality());
+
+            if (terminating)
+                setState(TelemetryState.terminating_push_in_progress);
+            else
+                setState(TelemetryState.push_in_progress);
+
+            return new PushTelemetryRequest.Builder(subscription.clientInstanceId(),
+                subscription.subscriptionId(),
+                terminating,
+                compressionType,
+                metricsData);
+        } else {
+            throw new IllegalTelemetryStateException(String.format("Cannot make telemetry request as telemetry is in state: %s", state()));
+        }
+    }
+
+    public static Set<MetricName> validateMetricNames(List<String> requestedMetrics) {
+        Set<MetricName> set;
+
+        if (requestedMetrics == null || requestedMetrics.isEmpty()) {
+            // no metrics
+            set = Collections.emptySet();
+        } else if (requestedMetrics.size() == 1 && requestedMetrics.get(0).isEmpty()) {
+            // TODO: TELEMETRY_TODO: determine the set of all metrics
+            set = new HashSet<>();
+        } else {
+            // TODO: TELEMETRY_TODO: prefix string match...
+            set = new HashSet<>();
+        }
+
+        return set;
+    }
+
+    public static List<CompressionType> validateAcceptedCompressionTypes(List<Byte> acceptedCompressionTypes) {
+        List<CompressionType> list = null;
+
+        if (acceptedCompressionTypes != null && !acceptedCompressionTypes.isEmpty()) {
+            list = new ArrayList<>();
+
+            for (Byte b : acceptedCompressionTypes) {
+                int compressionId = b.intValue();
+
+                try {
+                    CompressionType compressionType = CompressionType.forId(compressionId);
+                    list.add(compressionType);
+                } catch (IllegalArgumentException e) {
+                    log.warn("Accepted compression type with ID {} provided by broker is not a known compression type; ignoring", compressionId);
+                }
+            }
+        }
+
+        // If the set of accepted compression types provided by the server was empty or had
+        // nothing valid in it, let's just return a non-null list, and we'll just end up using
+        // no compression.
+        if (list == null || list.isEmpty())
+            list = Collections.emptyList();
+
+        return list;
+    }
+
+    public static Uuid validateClientInstanceId(Uuid clientInstanceId) {
+        return clientInstanceId.equals(Uuid.ZERO_UUID) ? Uuid.randomUuid() : clientInstanceId;
+    }
+
+    public static int validatePushInteravlMs(Integer pushIntervalMs) {
+        // TODO: TELEMETRY_TODO: I don't think we'll need this, but maybe it's good to be paranoid?
+        return pushIntervalMs != null && pushIntervalMs > 0 ? pushIntervalMs : 10000;
+    }
+
+    public static CompressionType preferredCompressionType(List<CompressionType> acceptedCompressionTypes) {
+        if (acceptedCompressionTypes != null && !acceptedCompressionTypes.isEmpty()) {
+            // Broker is providing the compression types in order of preference. Grab the
+            // first one.
+            return acceptedCompressionTypes.get(0);
+        } else {
+            return CompressionType.NONE;
+        }
+    }
+
+    public static MetricType metricType(KafkaMetric kafkaMetric) {
+        Measurable measurable = kafkaMetric.measurable();
+
+        if (measurable instanceof Gauge) {
+            return MetricType.gauge;
+        } else if (measurable instanceof Histogram) {
+            return MetricType.histogram;
+        } else if (measurable instanceof CumulativeSum) {
+            return MetricType.sum;
+        } else {
+            throw new InvalidMetricTypeException("Could not determine metric type from measurable type " + measurable + " of metric " + kafkaMetric);
+        }
+    }
+
+    public static ByteBuffer serialize(Collection<TelemetryMetric> telemetryMetrics,
+        CompressionType compressionType,
+        TelemetrySerializer telemetrySerializer) {
+
+        try {
+            try (ByteBufferOutputStream compressedOut = new ByteBufferOutputStream(1024)) {
+                try (OutputStream out = compressionType.wrapForOutput(compressedOut, RecordBatch.CURRENT_MAGIC_VALUE)) {
+                    telemetrySerializer.serialize(telemetryMetrics, out);
+                }
+
+                return (ByteBuffer) compressedOut.buffer().flip();
+            }
+        } catch (IOException e) {
+            throw new KafkaException(e);
+        }
+    }
+
+    public static TelemetryManagementInterface maybeCreate(AbstractConfig config, Time time, String clientId) {
+        boolean enableMetricsPush = config.getBoolean(CommonClientConfigs.ENABLE_METRICS_PUSH_CONFIG);
+        return maybeCreate(enableMetricsPush, time, clientId);
+    }
+
+    public static TelemetryManagementInterface maybeCreate(boolean enableMetricsPush, Time time, String clientId) {
+        if (enableMetricsPush)
+            return new TelemetryManagementInterface(time, clientId);
+        else
+            return null;
+    }
+
+    public static String convertErrorToReason(Throwable error) {
+        // TODO: TELEMETRY_TODO: properly convert the error to a "reason"
+        return String.valueOf(error);
+    }
+
+    public TelemetryState validateTransition(TelemetryState oldState, TelemetryState newState) {
+        switch (oldState) {
+            case initialized:
+                // If we're just starting up, the happy path is to next request a subscription.
+                //
+                // Don't forget, there are probably error cases for which we might transition
+                // straight to terminating without having done any "real" work.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.subscription_needed, TelemetryState.terminating);
+
+            case subscription_needed:
+                // If we need a subscription, the main thing we can do is request one.
+                //
+                // However, it's still possible that we don't get very far before terminating.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.subscription_in_progress, TelemetryState.terminating);
+
+            case subscription_in_progress:
+                // If we are finished awaiting our subscription, the most likely step is to next
+                // push the telemetry. But, it's possible for there to be no telemetry requested,
+                // at which point we would go back to waiting a bit before requesting the next
+                // subscription.
+                //
+                // As before, it's possible that we don't get our response before we have to
+                // terminate.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.push_needed, TelemetryState.subscription_needed, TelemetryState.terminating);
+
+            case push_needed:
+                // If we are transitioning out of this state, chances are that we are doing so
+                // because we want to push the telemetry. Alternatively, it's possible for the
+                // push to fail (network issues, the subscription might have changed, etc.),
+                // at which point we would again go back to waiting and requesting the next
+                // subscription.
+                //
+                // But guess what? Yep - it's possible that we don't get to push before we have
+                // to terminate.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.push_in_progress, TelemetryState.subscription_needed, TelemetryState.terminating);
+
+            case push_in_progress:
+                // If we are transitioning out of this state, I'm guessing it's because we
+                // did a successful push. We're going to want to sit tight before requesting
+                // our subscription.
+                //
+                // But it's also possible that the push failed (again: network issues, the
+                // subscription might have changed, etc.). We're not going to attempt to
+                // re-push, but rather, take a breather and wait to request the
+                // next subscription.
+                //
+                // So in either case, noting that we're now waiting for a subscription is OK.
+                //
+                // Again, it's possible that we don't get our response before we have to terminate.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.subscription_needed, TelemetryState.terminating);
+
+            case terminating:
+                // If we are moving out of this state, we are hopefully doing so because we're
+                // going to try to send our last push. Either that or we want to be fully
+                // terminated.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.terminated, TelemetryState.terminating_push_in_progress);
+
+            case terminating_push_in_progress:
+                // If we are done in this state, we should only be transitioning to fully
+                // terminated.
+                return validateTransitionAllowed(oldState, newState, TelemetryState.terminated);
+
+            case terminated:
+                // We should never be able to transition out of this state...
+                return validateTransitionAllowed(oldState, newState);
+
+            default:
+                // We shouldn't get to here unless we somehow accidentally try to transition out
+                // of the terminated state.
+                return validateTransitionAllowed(oldState, newState);
+        }
+    }
+
+    public static TelemetryState validateTransitionAllowed(TelemetryState oldState, TelemetryState newState, TelemetryState... allowableStates) {
+        if (allowableStates != null) {
+            for (TelemetryState allowableState : allowableStates) {
+                if (newState == allowableState)
+                    return newState;
+            }
+        }
+
+        throw new IllegalTelemetryStateException(String.format("Invalid transition from %s to %s", oldState, newState));
     }
 
 }
